requirements.txt
greenlet==2.0.2
python-dotenv==1.0.0
SQLAlchemy==2.0.18
typing_extensions==4.7.1
